feat(concurrency): Improve JDK 21 structured task scope instrumentation
Add instrumentation alias
Limit activation to Java < 25
Fix context store declaration and usage to match RunnableInstrumentation
Remove unused generic
